[ENH] [REFACTOR] imports to improve performance of the cli. (#281)
* [ENH] [REFACTOR] imports to improve performance of the cli.

* [FIX] type annotation.

* [ENH] lazy import to improve cli performance.

* [FIX] `sphinxarg.ext` release `0.5.0` breaks doc build.

Issue described at https://github.com/sphinx-doc/sphinx-argparse/issues/56

* [UPDATE] dependency requirement for sphinx-argparse.
