+++ .gitignore
@@ -37,3 +37,6 @@ env/
 # docs
 nipoppy_cli/docs/build
 nipoppy_cli/docs/source/schemas/*.json
+
+# OSX
+.DS_STORE
