+++ nipoppy_cli/nipoppy/workflows/dataset_init.py
@@ -10,8 +10,8 @@ from nipoppy.utils import (
     DPATH_TRACKER_CONFIGS,
     FPATH_SAMPLE_CONFIG,
     FPATH_SAMPLE_MANIFEST,
-    StrOrPathLike,
 )
+from nipoppy.env import StrOrPathLike
 from nipoppy.workflows.base import BaseWorkflow
