+++ nipoppy_cli/nipoppy/workflows/pipeline.py
@@ -11,6 +11,7 @@ from pathlib import Path
 from typing import Optional
 
 import bids
+from nipoppy.env import BIDS_SESSION_PREFIX, BIDS_SUBJECT_PREFIX
 from pydantic import ValidationError
 
 from nipoppy.config.boutiques import (
@@ -19,9 +20,6 @@ from nipoppy.config.boutiques import (
 )
 from nipoppy.config.pipeline import ProcPipelineConfig
 from nipoppy.utils import (
-    BIDS_SESSION_PREFIX,
-    BIDS_SUBJECT_PREFIX,
-    StrOrPathLike,
     add_pybids_ignore_patterns,
     check_participant_id,
     check_session_id,
@@ -32,6 +30,7 @@ from nipoppy.utils import (
     process_template_str,
     session_id_to_bids_session,
 )
+from nipoppy.env import StrOrPathLike
 from nipoppy.workflows.base import BaseWorkflow
