+++ nipoppy_cli/pyproject.toml
@@ -42,7 +42,7 @@ doc = [
     "furo>=2024.1.29",
     "pygments-csv-lexer>=0.1.3",
     "sphinx>=7.2.6",
-    "sphinx-argparse>=0.4.0",
+    "sphinx-argparse>=0.4.0, !=0.5.0",
     "sphinx-autoapi==3.0.0",
     "sphinx-copybutton>=0.5.2",
     "sphinx-jsonschema>=1.19.1",
